Replaced fireFuzzy.cpp with the improved/*functional* version.
As you can see, a rewrite made everything simpler! It doesn't crash now. For the most part, the sprite is "done", but there's one more thing I need to finish, as well as another thing I want to implement. Anyway, if you want to test it out, change all instances of "chorobon_flame" to just "chorobon".